refactor: Consolidate card types and styles into a new types file
